+++ src/main/java/refinedstorage/api/network/registry/INetworkRegistryProvider.java
@@ -1,0 +1,5 @@
+package refinedstorage.api.network.registry;
+
+public interface INetworkRegistryProvider {
+    INetworkRegistry provide(int dimension);
+}