+++ src/main/java/refinedstorage/block/BlockController.java
@@ -106,7 +106,7 @@ public class BlockController extends BlockBase {
     @Override
     public void breakBlock(World world, BlockPos pos, IBlockState state) {
         if (!world.isRemote) {
-            ((TileController) world.getTileEntity(pos)).getNodeGraph().disconnectAll();
+            ((TileController) world.getTileEntity(pos)).onDestroyed();
         }
 
         super.breakBlock(world, pos, state);