+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Refined Storage Changelog
 
+### 0.9.4
+- Fixed mod not working without JEI (raoulvdberge)
+
 ### 0.9.3
 - Updated German translation for Fluid Storage (0blu)
 - Updated Dutch translation for Fluid Storage (raoulvdberge)